refactor: 유틸 함수 더욱 세분화 해서 모듈화
