+++ packages/client/src/components/sidebar/RecentlyViewed.tsx
@@ -1,9 +1,9 @@
 import SidebarCoin from '@/components/sidebar/SidebarCoin';
 import { useSSETicker } from '@/hooks/useSSETicker';
-import { formatData } from '@/utility/formatData';
+import { formatData } from '@/utility/format/formatSSEData';
 import useRecentlyMarketStore from '@/store/recentlyViewed';
-import { convertToQueryString } from '@/utility/marketDataUtil';
 import { useRecentlyMarket } from '@/hooks/useRecentlyMarket';
+import { convertToQueryString } from '@/utility/queryString';
 
 function RecentlyViewed() {
 	const { recentlyViewedMarketList } = useRecentlyMarketStore();
@@ -20,9 +20,7 @@ function RecentlyViewed() {
 		<div className="flex flex-col h-full overflow-y-auto [&::-webkit-scrollbar]:hidden p-4">
 			<div>
 				<span className="text-lg font-semibold">최근 본 종목</span>
-				<span className="ml-2 text-xs text-gray-700">
-					10개 종목만 표시
-				</span>
+				<span className="ml-2 text-xs text-gray-700">10개 종목만 표시</span>
 			</div>
 			<div className="border border-solid border-gray-300 my-3"></div>
 			{recentlyViewedMarketList.length !== 0
