+++ packages/client/src/utility/format/formatSSEData.ts
