+++ packages/client/src/utility/queryString.ts
@@ -1,0 +1,3 @@
+export function convertToQueryString(marketList:string[]){
+	return marketList.map(market=>`market=${market}`).join('&')
+}
